+++ src/views/Post.vue
@@ -97,16 +97,24 @@ export default {
         })
     },
     deletcPost(index, post) {
-      ApiService.post.delete_post(post.id)
-      .then(data => {
-        console.log(data)
-        this.tableData.splice(index, 1)
-        this.$message.success('删除成功')
-      })
-      .catch(error => {
-        console.log(error)
-        this.$message.error('删除失败')
-      })
+      this.$confirm('确定要删除吗？ ', '提示', {
+        onfirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => { // 向请求服务端删除
+          ApiService.post.delete_post(post.id)
+          .then(data => {
+            console.log(data)
+            this.tableData.splice(index, 1)
+            this.$message.success('删除成功')
+           })
+          .catch(error => {
+            console.log(error)
+            this.$message.error('删除失败')
+           })
+      }).catch(() => {
+        this.$message.info('已取消删除!');
+      });
     },
     handleCommit(form) {
       if (this.currentEdit !== -1) {
