+++ src/views/User.vue
@@ -41,7 +41,7 @@ export default {
     return {
       map_type: {
         admin: "管理员",
-        user: "一般用户"
+        user: "普通用户"
       },
       tableData: [],
       pagination: {}
@@ -71,15 +71,24 @@ export default {
       });
     },
     deletcUser(index, row) {
-        ApiService.user.delete_user(row.id)
-        .then(data => {
+      this.$confirm('确定要删除吗？ ', '提示', {
+        onfirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => { // 向请求服务端删除
+          ApiService.user.delete_user(row.id)
+          .then(data => {
             console.log(data)
             this.tableData.splice(index, 1)
             this.$message.success('删除成功')
-        })
-        .catch(error => {
+          })
+          .catch(error => {
             this.$message.error('删除失败')
-        })
+          })
+      }).catch(() => {
+        this.$message.info('已取消删除!');
+      });
+
     }
   }
 };
